--- src/components/Join/Join.css
@@ -1,81 +1,0 @@
-/* Join.css */
-.join-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-  background-color: #f0f2f5;
-}
-
-.join-content {
-  background: white;
-  padding: 40px;
-  border-radius: 8px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  width: 100%;
-  max-width: 400px;
-}
-
-.join-title {
-  text-align: center;
-  margin-bottom: 30px;
-  font-size: 24px;
-  color: #333;
-  border-bottom: 2px solid #333;
-  padding-bottom: 10px;
-}
-
-.input-group {
-  margin-bottom: 20px;
-}
-
-.input-group label {
-  display: block;
-  margin-bottom: 8px;
-  font-size: 14px;
-  color: #555;
-}
-
-.input-group input {
-  width: 100%;
-  padding: 12px;
-  border: none;
-  border-radius: 6px;
-  background-color: #f5f5f5;
-  font-size: 14px;
-}
-
-.input-group input.error {
-  border: 2px solid #ff4444;
-}
-
-.error-msg {
-  display: block;
-  margin-top: 5px;
-  color: #ff4444;
-  font-size: 12px;
-}
-
-.join-button {
-  width: 100%;
-  padding: 14px;
-  background-color: #5b73e8;
-  color: white;
-  border: none;
-  border-radius: 6px;
-  font-size: 16px;
-  cursor: pointer;
-  margin-top: 20px;
-  transition: background-color 0.3s;
-}
-
-.join-button:hover {
-  background-color: #4a61d1;
-}
-
-@media (max-width: 480px) {
-  .join-content {
-    padding: 30px 20px;
-    margin: 0 15px;
-  }
-}
